+++ .gitignore
@@ -8,3 +8,4 @@ __pycache__
 /dist
 /staramr.egg-info
 /staramr/databases/data/
+/.eggs
