+++ staramr/databases/dr/phenotype/ARGDrugTable.py
@@ -1,7 +1,9 @@
 import pandas
-
+from os import path
 
 class ARGDrugTable:
+
+    DEFAULT_DATA_DIR=path.join(path.dirname(__file__), '..', 'data')
 
     def __init__(self, file):
         self.file = file
